* commit '027624f6ce76439f05f9b14959ecaf748c47d30f':
  fix entitites functions
  add removePostByUser function
  added additional check to entity
  add email checking to user entity
  fix createUser function
  add default error-messages
  fix User model import
  remove errorThrow function from errorService
  remove try..catch from routes
  add errors handle
  add errors handle
  add errors handle
  fix error handle
  fix error handlers
  fix error handle
  update validate.js
  fix: error hanle works
  fix: errors hanle
  add customn error hanles

+++ src/entities/category.js
@@ -1,23 +1,44 @@
 const Category = require("../models/category");
+const PostEntity = require("./post");
+const Error404 = require("../middleware/error/error404");
+const Error400 = require("../middleware/error/error400");
 
 async function getAllCategories() {
   return Category.query().select();
 }
 
 async function getCategoryById(id) {
+  const category = await Category.query().findById(id);
+  if (!category) {
+    throw new Error404();
+  }
   return Category.query().findById(id);
 }
 
-async function createCategory(post) {
-  return Category.query().insert(post);
+async function createCategory(category) {
+  await Category.query().insert(category);
 }
 
 async function updateCategory(id, data) {
-  return Category.query().findById(id).patch(data);
+  const category = await Category.query().findById(id);
+  if (!category) {
+    throw new Error404();
+  }
+  await Category.query().findById(id).patch(data);
 }
 
 async function removeCategoryById(id) {
-  return Category.query().findById(id).delete();
+  const category = await Category.query().findById(id);
+  const postsByCategory = await PostEntity.getPostsByCategory(id);
+  if (!category) {
+    throw new Error404();
+  }
+  if (postsByCategory !== 0) {
+    throw new Error400(
+      "You cannot remove the category because posts of this category exist"
+    );
+  }
+  await Category.query().findById(id).delete();
 }
 
 module.exports = {
